Refactoring Main into new class Game
This refactoring allows me to bring large amounts of data into the main
for use outside the main using static calls. This keeps me from having
to pass instance variables throughout the constructor structure and
making a real mess of method calls that go between modules.
